requirements (1).txt
openai-whisper
streamlit
numpy
torch
pydub
ffmpeg
gtts
